[StentExp] FIX 2 bugs entering the plastic phase
2 problems are fixed regarding the passage from an elastic point to a
plastic point :
- The test of the yield function that preceded the entering into plastic
phase was made comparing to 0. An arbitrary threshold is fixed instead
(and is to be refined)
- If some of the points were already deforming plastically and an
elastic point was detected to switch to plastic, the unrealistic elastic
stress was not corrected immediately. This led to a point switching from
plastic to elastic and elastic to plastic every two time steps. The fix
is simply to compute incrementally a new plastic stress once the stress
is detected as plastic.