[StentExp] FIX bug getting out of plastic state
The test made on the (plastic) stress points to test if they go back to
an elastic phase was made after the stress increment Newton-Raphson
iterations (after all the points had been updated). This caused the
iterations to actually bring a point which had returned to elastic, back
to the plastic state (which doesn't make sense).

To prevent this from happening, we now test each plastic point before
the beginning of the iterations, basing on the elastic predictor. The
update of the point mechanical states is thus made in the
computeStressIncrement method, rather than in accumulateNonLinearForces.

For the moment the code is not working properly: going into a
post-plastic state would require to store the accumulated plastic
strain in memory, which is currently not the case. This is to be handled
in the same place as the post-plastic check, i.e. in the
computeStressIncrement method.