[StentExp] ADD handling of post-plastic phase
When a point is detected to go back from plastic state to elastic state,
we now compute the total plastic strain that occured during the plastic
deformation and store it. In parallel the point status is changed to
POSTPLASTIC.
When in POSTPLASTIC state, the stresses at a point are computed
elastically, after having substracted the plastic part of the strain
from the total strain.
To implement properly this new handling, the method dedicated to compute
stress increments now returns directly the new resulting stress point.
The accumulateNonLinearForces method simply calls this routine and uses
the new stress point as it is.

We also corrected a mistake in the material behaviour matrix (C), where
the shear strains were computed as normal shear strains, instead of
engineering shear strains (as expected from the Be matrices
expressions). The corresponding '2' factor has been deleted.

Finally the inverse of this material behaviour matrix is now explicitly
computed rather than deduced through QR Eigen decomposition.

NB: for the moment, the plastic beam returns to the rest position even
after plastic deformation occured. This has to be investigated and could
be due to the distinction between plastic and elastic Gauss points, or
to the perfectly plastic model implemented here (with no hardening)