[StentExp] FIX bug in plastic deformation
The points global DoFs were saved in variable _lastPos after the stress
increment iterations were conducted on each element. If there were more
than 1 beam element, this would cause the computed displacement
increment to be zero for all the elements except the first one.
To fix this, the update on _lastPos is now made at the end of the
addForce method.