[StentExp] MODIFY computation of plastic strain
The computation of the plastic strain at the end of the plastic phase
is not rigorous, judging by the incremental aspect of the plastic
computation. A more proper way is to compute the plastic strain
increments at each time step : once we reach the last stress point of
the plastic deformation, these increments should be integrated along the
deformation path (in some way) to obtain the total plastic strain.

For the moment, the increments are just summed, and the proper
integration process is still to be determined. We can note that summing
the increments gives the same result as the system solution we were
using. This is due to the linear behaviour of the (perfectly) plastic
deformation. It should not be the case if hardening was implemented.

NB: the out-of-plastic-state threshold is modified, but should be
computed relatively to the formerly computed values