[StentExp] SUPPRESS yield stress general field
The general yield stress parameter used in the computation of the Von
Mises yield function is changed for local yield stress parameters which
are stored in the BeamInfo object. These local elements prepare the
implementation of an isotropic hardening behaviour, according which the
yield stress becomes a hardening parameter that changes with plastic
deformation.
For the moment, this parameter is constant, which gives the same
behaviour (without hardening) as before.