[SofaPython] ADD systematically sys.argv in PythonEnvironment
The current behavior was to not set a sys.argv if there where not argument passed in
the Sofa command line. The consequence was that in the python code it was needed
to do
if hasattr(sys, "argv"):
   ...

Which is rather uncommon.

This Commit change this behavior by setting systematically sys.argv and additionally fix typos (ending white spaces)