[StentExp] PREPARE a different local frame handling
We add a new structure _beamLocalOrientations to the MultiBeamForceField
in order to test a different approach for handling the element local
frames.
Until now, these were computed directly from the last 3 DoFs of each
element first node (node a). They were then stored in the 'quat' field
of the BeamInfo strcuture.

The problem with this implementation is that it prevents one to connect
2 or more beam elements through their first node. Indeed if 2 beams were
to share the same first node, the orientation of this node could not
properly describe the local frame orientation of both elements.
The idea here is to store the local frame orientation not through the
nodes (i.e. the DoFs) themselves, but directly trhough each beam
element. Hence the creation of the _beamLocalOrientations field.

NB1: for the moment, we created a Data structure
(_inputLocalOrientations) to manually set up these local frame
orientations. In the future, we will investigate how to automatically
determine the orientations from the topology.

NB2: the newly created _beamLocalOrientations field could also have been
implemented at the BeamInfo level, just as the current 'quat' variable
is. This is done on purpose to keep both implementation active as long
as the new one remains a prototype. If it works, any redundancy should
be suppressed.

NB3: a minor modification was made to variable _UTS (Ultimate Tensile
Strength) to rename it properly (_yieldStress)