[StentExp] ADD bbox handling to MultiBeam component
Several visualisation issues in the viewer (inappropriate zoom, fixed
scene centre, impossible to move the camera, ...) were due to the lack
of bounding box handling at the level of the MultiBeamForceField
component.
We add this handling through a method computeBBox, which is the same as
in the BeamFEMForceField component. The computation of the bounding box
is based on the node positions of the connected beam elements.