[StentExp] CHANGE yield stress parameter definition
The yield stress (which is used in the von mises stress function) is now
specified the same way as the other mechanical parameters. It is
therefore possible to specify it as an argument when declaring the
component from a SOFA scene file. The parameter itself is then stored at
the BeamInfo structure level.