[StentExp] CHANGE the beam element display
We moved from a quadrilateral rigid display to a new display which makes
use of precomputed shape functions.
The idea is to discretise the centreline of the beam into a certain
number of segments, which are then tracked the same way as the Gauss
points. The number of segments is given by the _nbCentrelineSeg field in
MultiBeamForceField.h. As the two extremity point positions are known,
we have to track (_nbCentrelineSeg - 2) inside points.
In order to do so, we precompute (_nbCentrelineSeg - 2) shape functions
from the coordinates of the corresponding centreline points. These shape
functions are stored in the _drawN field of the BeamInfo structure, and
used in the drawElements method, in the same way we use the _N shape
functions to track the Gauss points.