[modules] ADD a new module name SofaEditor.
The aim of this module is to allows the editors build on top of Sofa
(eg runSofa, runSofa2) to communicate their state to components likes
Controller, PythonScriptController or (maybe one day
JavaScript controller).

The implementation is yet very basic but contains unittes as well as an
independent python binding.

PS: You may need cython to compile the python binding.