[StentExp] CORRECT stiffness matrix computation
We changed what seemed to be an error in the shape fucntions formulas we
used for the beam elements. The (supposed) mistake was related to an
additional '-' sign interfering in the computation of the displacement
along z resulting from a twisting moment (moment around the x axis).

As the shape functions expression is directly involved in the
computation of the Be matrices, we changed the corresponding expression
in the MultiBeamForceField component.

This change is coherent with the SOFA implementation of the beam
elements, as the lines of the stiffness matrix corresponding to torsion
give similar results in both implementations.