[StentExp] ADD Timoshenko shape functions
We implement shape functions compatible with the Timoshenko beam theory
in the MultiBeamForceField component. The expressions we use are taken
from Bazoune et al 2003 article: "Shape Functions of Three-Dimensional
Timoshenko Beam Element".

The change of shape functions has also consequences on the derivatives
of these functions, which are contained in matrix Be. We therefore
implement as well a second version of the Be matrix, corresponding to
Timoshenko beam theory. Detail of the computation is written in an
attached note.