[SofaPython3] Replace the use of Node::create with DAGNode::create in SceneLoader3.
The problem with Node::create is that it use the instance of
Simulation that is loaded (tree vs graph). Which is not really
consistant. It is much better that our API only work with DAG.
And if needed can be mapped to tree structure.

Eg:
  from Sofa.simulation import Node it is actually sub-loading
Sofa.Simulation.DAG.Node


This also allow to write explicit call as like this:
  from Sofa.Simulation.DAG import Node
vs
  from Sofa.Simulation.Tree import Node