[SofaPython3] FIX several problems and memory leaks in how we are caching data containers.
The general problem was that the caching system was holding reference
to destroyed object. So it was necessary to hold a smart pointer to the
data sources as long as they are in the cache.

Open question: do we really need such cache instead of storing the data in
a different object ?

Signed-off-by: Damien Marchal <damien.marchal@univ-lille1.fr>