[SofaPython3] ADD the wrap-around types to decorate data with their right type.
Currently the data are exposed as "DataContainer" in python. The reason is that it
allow to offer an unified way to manipulate all the data fields as well as insuring
that the beginEdit/endEdit pair are correctly handled. The user can also get
read or write access on the data field through numpy.view of the data.

Now, it is often convenient to manipulate our data with more concrete type (eg: Vec4,
Rigid, Frame, RGBAColor). One possibility would be to implement those data types
in the binding, the approach I tried in this commit is different as the concrete
types are implemented as a kind of façade over the numpy.array exposed by the DataContainer.
This is what is implemented in this commit.

The general idea is to wrap subtypes around a compatible DataContainer.
For the user point of view this look like that:
color = RGBAColor(obj.showColor)
v = Vec4(obj.position)
rigid = Rigid(obj.position)

Then it become possible to manipulate the DataContainer as if it
is a color (so using dedicated methods).
color.r()
color.hsv()

The previous syntax is in fact equivalent to write:
RGBAColor(obj.showColor.toarray())

So it is clear that because of the use of the toarray() function we are operating
in read-only mode. So do things like that:
color.r = 40 or color.setFromHSV(1,2,3)
will generate an exception.

In order to get write access it is needed to use the provided context manager in
the following way:
Eg:
with obj.showColor.writeable(RGBAColor) as color:
	color.r = 40

when the context left, the color object is still accessible but it returns back to
a read-only state.