[SofaKernel] Remove SOFA_SUPPORT_MAPPED_MASS.
The mapped masses does not seems to be used at all since years.
I think it is dating from SOFA_DEV. There is no way to
activate/deactivate it from CMakeLists.txt.

In case this is still an active topic. Please open an issue/PR in case
this is still usefull to you.