[Topology] Change: add a boolean data (d_propagateToDOF) to allow or not the propagation of the topological changes to the mechanical state immediately as soon as they occure.
This modification should make the concurrent execution of the topological changes and the simulation safe.
Default behavior (d_propagateToDOF=true)  the topological changes are propagated immediately.
If the boolean is set to false the topological state changes update are not propagated immediately and they can executed in another step of the simulaiton loop.