[StentExp] SUPPRESS method initLarge
We suppress the method initLarge defined and used in the
MultiBeamForceField component. This method was originally defined in the
BeamFEMForceField and was copied identically at the creation of the
MultiBeamForceField.
The purpose of initLarge is apparently to modify the quaternion
associated to each beam element as a field of the BeamInfo structure
(BeamInfo.quat). The method actually computes an intermediate
orientation between the orientation of node A and B (start and end) of
the beam element. If the orientations are the same, the beam quaternion
is set by default to the quaternion of node A.

However, it appears that, after the initialisation is done, each time
the beam quaternion is used (namely in methods addForce, addDForce and
addKToMatrix), it is systematically modified beforehand to be equal to
the quaternion of node A (in the beginning of the accumulateForceLarge
method). Consequently, the computation made in initLarge seems to be
ineffective.

A quick verification in Debug mode on SimpleTestScene.py suggets that
the 'quat' field of the BeamInfo structure is indeed not modified by any
other method than the ones mentionned above. The behaviour of the mesh
in the same test file also seems to be the same before and after the
removal of initLarge.
Although this is not a thorought confirmation that initLarge is
definitely not used, we adopt the modification for the moment.