[StentExp] REORGANISE redundant code
The modified code corresponds to the mechanical behaviour computation
during the first elastic phase of the deformation.
Part of the code was unnecessarily duplicated due to the multiplication
of cases (elasticity, Krabbenhoft plasticity, Muller plasticity). We
suppressed any redundance between the different scenarios.

/!\ The case of Muller plasticity corresponds to a first try in
implementing a plastic behaviour. It is however irrelevant and should be
deprecated.