[StentExp] FIX minor error in post-plastic behaviour
The post-plastic behaviour consists of an elastic deformation which
takes into account the plastic history of the material. The computation
of the elastic part of the stress relied on an irrelevant displacement
value, computed one time step ago.
The correction did not show (qualitatively) any important change on the
simulation regarding the test scenarios.