[StentExp] MODIFY plastic stress increment computation
The first step of the implicit method used in the computation of plastic
stress increments involves the resolution of a simple system, which can
be solved analytically. The only conditon on this system is the use of
specific initial values for the unknowns (cf Krabbenhoft's lecture
notes).
We implement this analytical solution, in order to enter the iterations
only if the first increment found is not acceptable (i.e. if yield
function computed at the new stress point is too far from zero). Doing
so, we save the resolution of a 7x7 system for each beam element, at
each time step.

NB: when using the Von-Mises yield function, the system from which we
compute the stress increments actually admits this first solution as a
closed-form (because of the regularity of the function).

At the moment, the analytical solution is implemented in addition to the
implicit resolution algorithm, for verification purposes.
In the long term, the first step of the implicit iterations should be
replace by this analytical solution, followed by iterations if
necessary (which technically should not be the case as long as Von-Mises
function is used).