[StentExp] MODIFY the stress increment computation
Following previous commit 81316b9fea92dec3004c432db2fce71982cb90e4, we
implement the analytical solution of the first iteration in the
implicit procedure used to compute stress increments.

NB: while using the Von Mises yield function, the second part of the
algorithm (effectively implementing the implicit resolution) should
actually not be used. This comes from the fact that the first step
solution is in this case correct, because of the yield function
regularity.