[StentExp] ADD parameter in MultiBeamFF creation
This commits adds a boolean parameter in the creation of a
MultiBeamForceField object, indicating if the corresponding beam
elements should have a perfectly plastic behaviour.
At the moment, only the additional option is implemented, not the
perfectly plastic behaviour itself.

NB: Only the simple scenario tests and two more complete test scenarios
implement this possibility.