[StentExp] ADD perfectly plastic behaviour
Following the previous commit, we implement an optional perfectly
plastic behaviour for the beam elements. This behaviour is used only on
Gauss points which are in plastic state.

/!\ First tests on simple scenarios show a coherent plastic behaviour
(the strain keeps being increased as long as external forces remain
constant). However when external forces are reduced, the beam element
doesn't enter into a post-plastic (elastic) deformation phase, as it
should. This bug is likely to be related to the plastic state ending
condition used at the moment, and must be addressed in further commits.