[StentExp] ADD 1D plastic constitutive law model
To prepare the implementation of (isotropic) hardening, we need to be
able to compute the tangent modulus at eah time step. This requires the
derivation of the 1D plastic (non-linear) constitutive law associated to
the simulated material (in our case, for instance, the L605
cobalt-chromium alloy).
As the exact expression of the relation between stress and strain is
higly specific, the idea is to use a classic model describing the 1D
stress-strain curve, and fit the parameters of this model on
experimental data. The first model we try is the Ramberg-Osgood model.

From the implementation point of view, we define a virtual class
PlasticConstitutiveLaw, from which should inherit every concrete model
to be tested. As a first example, we implement the Ramberg-Osgood model
in this way.

An object of type PlasticConstitutiveLaw should then be used in the
Force Field to implement a hardening behaviour. This should be the
purpose of the next commit.