[StentExp] CORRECT dot product error in plastic computation
The computations we carry out during the plastic phase involve dot
products between tensor quantities, notably in:
  - the implicit computation of stress increments with the radial return
algorithm,
  - the tangent stiffness update,
  - the condition to go from plastic to post-plastic phase.

A (serious) problem is that the objects we use are implemented using the
Voigt notation, meaning that half of their non-diagonal elements is not
explicitely represented. The consequence is that the dot product we use
at the moment is *underestimated*, as these elements are not part of the
computation.

To solve this issue, we implement a dedicated method voigtDotProduct
which computes the dot product between second order tensors written in
Voigt notation. For the moment, it replaces the former computations in
compteStressIncrement and updateTangentStiffness.

NB: it should be checked whether the use of the Voigt notation is
problematic in the rest of the computation.