[StentExp] CORRECT out of plastic phase condition
Consequently to the previous commit, we also replace the computation
made in the staysInPlastic method, to use the voigtDotProduct formula.
The name of the method was also changed to something more specific
(goToPostPlastic, as opposed to goToPlastic).

We also implemented a specific method to compute the Von Mises
equivalent stress, used in the computation of the yield criterion.