[StentExp] MODIFY perfect plasticity model
We replace the 'genuine' implementation of perfect plasticity (directly
derived from the 1D example) by a more precise version described in
'Theoritical foundation for large scale computations for nonlinear
material behavior', Hugues et al, 1984.
An important change, as already introduced in the last 3 commits, is the
use of the deviatoric stress tensor to compute the normal to the yield
surface, instead of the analytical expression of the gradient.