[StentExp] REMOVE _partial_segment_list feature
We remove a feature initially taken from the BeamFEMForceField
component, allowing to mechanically deform only part of the beam
elements (specified by a list in a dedicated field).
This feature is a priori of no use in our case, and it's removal
makes code readability better. Otherwise, it can be retrieved from the
BeamFEMForceField code