[SofaGeneralObjectInteraction] Remove component_state in AttachConstraint
In the current state there is no need for ComponentState as the "error" message
are in fact "warning/info" because the detected problems are not preventing the
object to function properly.

I did various minor cleaning (to remove compiler warnings).
There is still a lot of uglyness in this class if a brave hacker found it,
don't hesitate.