[StentExp] CHANGE boolean field for plastic info
Until now, we used a unique boolean variable _isDeformingPlastically to
adapt the mechanical computation. This variable was common to all beam
elements contained in the MultiBeamForceField component, leading
consequently to a global "binary" behaviour:
  - if all the Gauss points of all the beam elements were in ELASTIC
state, then the computation of internal forces would be done directly
from the precomputed elastic stiffness matrix (as in BeamFEMForceField).
  - If at least one point was in PLASTIC state, the computation would be
done incrementally for all points, in all elements. Concretely, the
incremental resolution would take place for the computation of the new
stress state, from which internal forces would be retrieved by reduced
integration. For fully elastic elements, this apparently didn't change
the computation as stress increments were computed elastically for each
ELASTIC Gauss point, and the computation of the (tangent) stiffness
matrix used the elatsic constitutive law (C) instead of the
elasto-plastic law (Cep). Nevertheless, this lead to an unnecessary
recomputation of the elastic stiffness at each time step, for each fully
elastic beam element.

This commit aims at defining a beam specific variable to indicate
whether at least one of the element Gauss point is in PLASTIC state.
This should lead to more efficient computation, as fully elastic
elements could be handled separately from partially plastic ones.

/!\ Another issue that is not addressed by the present commit is the
computation in the post-plastic phase. Until now, POST-PLASTIC Gauss
points were handled incrementally as well, in the same way ELASTIC
points were during plastic deformation. It seems that an
ELASTIC/POST-PLASTIC beam element was handled correctly, as the tangent
stiffness computed at each time step was equal to the elastic stiffness,
but once again, the computation cost was unnecessarily high.
Right now, we consider that a fully ELASTIC/POST-PLASTIC beam is
"elastic" (i.e. _isPlasticBeam = false). In the current code, this makes
such an element being handled through the accumulateForceLarge method,
instead of the accumulateNonLinearForce method. In this case the
computation is **wrong** (!) as the plastic history is not considered
for POST-PLASTIC points in elastic beam elements.

A solution to this would be to fusion the accumulateForceLarge and
accumulateNonLinearForce methods, and check for each element if the
tangent stiffness matrix has to be computed or not. This should be the
purpose of the next commits.