[StentExp] REVISE the plasticity model implementation
We procceed to a major revision in the way plasticity is handled in the
MultiBeamForceField component. The new implementation mostly modifies
the behaviour of the addForce method. The main modifications are
described below.

1) The global variable _isDeformingPlastically, common to all beam
elements of the component, is replaced by an element-specific variable
(beamMechanicalState), describing the mechanical state of the element's
Gauss points. We switch the variable type from boolean to
MechanicalState, with three possible values: ELASTIC, PLASTIC, and
POSTPLASTIC. In the case of a whole beam element, the values are
interpreted as follows :
  - ELASTIC means that the mechanical state of all the element's Gauss
points is ELASTIC,
  - PLASTIC means that the mechanical state of at least one Gauss point
is PLASTIC,
  - POSTPLASTIC indicates that the points' mechanical state is either
ELASTIC or POSTPLASTIC (but not PLASTIC).
The whole plastic computation is modified accordingly, so that the beam
elements' mechanical state is correctly updated when new stress tensors
are computed at each time step.

2) The global process triggered by a call to addForce is simplified.
Until now, a call to addForce would result in a call to an elastic
solution (accumulateForceLarge) if the global variable
_isDeformingPlastically was evaluated as false, and in a call to an
incremental plastic solution (accumulateNonLinearForce) if the variable
was true. In the second case, all beam elements were handled plastically
as long as at least one Gauss point in *one* element was PLASTIC. This
resulted in a systematic recomputation of the tangent stiffness matrix,
even for ELASTIC and POST-PLASTIC elements, in which the computation
lead to the elastic stiffness matrix (and thus was useless). It also
resulted in reduced integration to compute internal forces in fully
elastic elements, for which a simple matrix computation (f = Ku) could
be done.
With this commit, we remove the accumulateForceLarge method and limit
the addForce method to a call to accumulateNonLinearForce. In this last
method, we call one of 3 new methods: computeElasticForce,
computePlasticForce, and computePostPlasticForce. Each one of them is in
charge of computing the nodal internal forces (locally) for each beam
element, depending on its mechanical state.

3) addKToMatrix is modified so that the elastic stiffness matrix is used
for beam elements in ELASTIC or POSTPLASTIC state, and the tangent
stifness matrix is used for PLASTIC elements.