[StentExp] CORRECT Von Mises gradient expression
Until now, the expression we used for the Von Mises yield gradient was
incorrect (due to calculation error). We correct the factor 6 associated
to non-diagonal elements by a correct factor 3.
We additionnally remove the yield stress from the method definition, as
Von Mises gradient does not depend on it (the use of the yield stress
until now was superfluous, and resulted from an unnecessary complex
expression of the gradient).