[StentExp] FUSION redundant code
In the same way we transfered the code of accumulateForceLarge into
accumulateNonLinearForce, we integrate the code of applyStiffnessLarge
into applyNonLinearStiffness.
Now, the computation of K*v in the right-hand side of the mechanical
system is entirely handled in applyNonLinearStiffness, and the
expression of the matrix K depends on the mechanical state of the beam
elements.