[StentExp] CLEAN deprecated code
We remove from the MultiBeamForceField component deprecated methods and
fields, introduced in preliminary incorrect implementations of the
plasticity model.

The removed methods are:
  - peudo_determinant_for_coef (*pseudo* inherited from BeamFEMForceField)
  - computePlasticForces
  - updatePlasticStrain

The removed types (typedefs) are:
  - VoigtTensor (redundant with VoigtTensor2)
  - elementPlasticStrain

The removed fields are:
  - _VDPlasticStrains (redundant with _plasticStrainHistory)
  - _VDPlasticCreep
  - _VDPlasticYieldThreshold

We note that most of these elements are redundant with the current
implementation of our plasticity model. Another example is the
_VDPlasticYieldThreshold, which is now comuted with the Von Mises yield
criterion.