[StentExp] REMOVE major deprecated code
We remove code corresponding to what we called the Virtual Forces
method, and the Muller plasticity model. The implementation of these
models corresponded to a very early vision of the plasticity model, and
was not based on a rigorous fundamental basis.
As it has not been used for the past year, we remove the corresponding
parts of the code, to improve code readability.

The removed methods are:
  - initPlasticityMatrix (used only in the init method)
  - updatePlasticityMatrix (used only in updatePlasticity)
  - updatePlasticity (not used anywhere)
  - totalStrainEvaluation (not used anywhere)

The removed types (typedefs) are:
  - VFStrain
  - plasticLimits
  - completePlasticZones
(all of them only used in the definition of related fields)

The removed fields are:
  - _VFPlasticYieldThreshold (not used in the implementation)
  - _VFPlasticMaxThreshold (not used in the implementation)
  - _VFPlasticCreep (not used in the implementation)
  - _VFPlasticStrains (only initialised)
  - _VFTotalStrains (only initialised)
  - _nodalForces (only initialised)
  - _plasticZones (only initialised)
  - _isPlasticZoneComplete (only initialised)
  - _M_loc (only initialised)

We consequently removed the _isPlasticMuller boolean data associated to
the component.
/!\ Because of this, most test files we are currently using will trigger
an error at execution time, for assigning a value to a non-existing
data. This issue should be addressed in the next commit