[StentExp] CORRECT elastic to plastic transition
A minor incoherence remained in the code regarding the transition from
an ELASTIC state to PLASTIC, for a beam element. The issue was that when
a Gauss point was detected as entering a PLASTIC state, its mechanical
state was changed to PLASTIC *before* the call to computePlasticForce
was made. Consequently, a test was carried out on the same point in
computePlasticForce to determine whether it stayed PLASTIC or became
POSTPLASTIC. Such a test would make no sense for a point in transition
between ELASTIC and PLASTIC as the previous stress tensor for the same
Gauss point lied inside the yield surface (elastic), rather than on the
yield surface (plastic).

This bug was corrected by stopping the computation in
computeElasticForce as soon as a Gauss point is detected as newly
PLASTIC. A call to computePlasticForce is made right after, where the
computation is carried out incrementally. This changes nothing for the
points which would have remained in ELASTIC state (the incremental
computation of the elastic predictor for these points will lead to the
same new stress as in computeElasticForce). On the other hand, newly
PLASTIC points will be handled correctly.
Internal forces are computed entirely in computePlasticForce, in which
the mechanical state of the beam and the Gauss points is properly
updated.
In addition to solving the issue mentionned above, this allows to leave
only elastic computation in computeElasticForce.

NB: we verified manually that the computation for ELASTIC points didn't
change, on SimpleTestScene.py (scenarioNb = 1). The incremental
computation gives the same new stresses for all ELASTIC Gauss points, up
to the last 3 decimals in double precision.