[StentExp] CORRECT bug in beam mechanical state handling
Following the recent changes in the implementation, the beam elements'
mechanical state was not updated correctly during the transitions from
ELASTIC to PLASTIC and POSTPLASTIC to PLASTIC. The fix is simply to
manually change the beam mechanical state when the first call to
computePlasticForce is made during the transition.

NB: with this correction, the transitions between mechanical states are
properly handled when the state of the element Gauss points changes.
However, we observe on SimpleTestScene.py (scenario 1) a "jump" of the
element at the end of the PLASTIC phase. Gauss points' mechanical state
undergoes a transition PLASTIC -> POSTPLASTIC -> PLASTIC -> POSTPLASTIC
before stabilising into POSTPLASTIC, which is illogical. This phenomenon
should be investigated in further commits