[StentExp] FIX bug in the perfect plasticity model
The perfectly plastic model, implemented from T.J.R. Hugues et al, 1984,
was incorrectly reproduced in the component code. The elastic predictor
was used in the computation of the plastic strain increment, instead of
the elastic increment. This caused the plastic strain increment to be
critically overestimated at each time stent, resulting in a strong
discontinuity when passing from PLASTIC to POSTPLASTIC.