[StentExp] FIX computation error due to Voigt notation
An important approximation was made in different parts of the mechanical
computation, because of a wrong handling of the contracted Voigt
notation for tensors.
Although the use of Voigt notation is relevant is some aspects
(typically to save memory space and prevent from carrying out useless
computation on symmetrical elements), it leads to uncomplete computation
whenever the resulting quantities are not tensors themselves.
For instance:
  - stress = C*strain is correct in Voigt notation, given that the
computed non-diagonal elements in the (Voigt) resulting stress tensor
are correctly evaluated.
  - (Be^T)*stress or (Be^T)*C*Be is incorrect, as missing rows in Be
lead to an incomplete summation in the computation. This is more obvious
by writing the 2nd-order tensors as concatenated (9x1) vectors and
4th-order tensors as concatenated (9x9) matrices, and by proceeding to
block computation.

In order to take into account the missing blocks in the computation, we
implement two additional methods, specifically dedicated to compute
(Be^T)*stress and (Be^T)*C*Be.

NB: After basic testing, it appears that the actual difference in the
computed values before and after correction is relatively low, as the
modifications are only associated to the non-diagonal elements of the
strain and stress tensors, whose intensity is low compared to the
diagonal elements' (due to our shape functions).

/!\ Additional investigation should be carried out to see if such
errors remain in the code, in other part involving tensors in Voigt
notation