[StentExp] CORRECT plastic to post-plastic transition
Until now, when a point was detected as going from a PLASTIC mechanical
state to POST-PLASTIC, the associated new stress tensor was taken as the
elastic predictor, without taking into account the plastic history.
We modify the computation made in the first time step of a POST-PLASTIC
Gauss point, so that the plastic history is correctly taken.