[SofaKernel] FIX bug in MeshTopology
In issue 746 is reported a bug when trying trying to simulate
scenes while the underlying topology is containing empty data set.

I narrowed the problem to be in LocalMinDistance::computeIntersection.
In this function we access the trianglesAroundVertex from a MeshTopology.

with code like that:
```cpp
 if (m_edgesInTetrahedron.empty() || i > m_edgesInTetrahedron.size()-1)
         createEdgesInTetrahedronArray();
    return m_edgesInTetrahedron[i];
```

While if there is not "nothing" to create because there is no edges/triangles
it access in there is still a return of m_edgesInTetrahedron[i] with indice = 0

To avoid that I change the code for:
```cpp
  const MeshTopology::EdgesInTetrahedron& MeshTopology::getEdgesInTetrahedron(TetraID i)
 {
     if (m_edgesInTetrahedron.empty() || i > m_edgesInTetrahedron.size()-1)
         createEdgesInTetrahedronArray();
    return m_edgesInTetrahedron[i];

    if (i < m_edgesInTetrahedron.size())
        return m_edgesInTetrahedron[i];

    return InvalidEdgesInTetrahedron;
}
```

So if there is nothing created we returns a valid object indicating either an
empty or one filled with invalid values.