[SofaPython] FIX C casts removing constness in Binding_Data
The problem is that it important when writing in data field to properly
notify the data successors that a data has changed. This is why write
access *must* be done using beginEdit/endEdit and abolutely not
by removing the const qualifier from q getValueVoidPtr using a cast.

The PR fix that.