[runSofa] CHANGE how QSpinBox handle mouse wheel.
When the mouse is over the "size" field of vector data fields
it start changing the value while on top of other field is scroll up/down
the view.

This inconsistant behavior is fixed in this PR. For doing that only the
selected spinbox can have their value changed.

For more details:
https://stackoverflow.com/questions/5821802/qspinbox-inside-a-qscrollarea-how-to-prevent-spin-box-from-stealing-focus-when