[SofaKernel] BaseData are set to non persistant by default.
This is a long standing issue in Sofa about not being able to save 'usable' scene in runSofa. Several PR had been made a long time ago.

There is multiple problems among which:
- keywords that are not Data field but parsing hook (eg: src, compatbility hook) needs to be converted into Data
- the Data are not persistant or not depending on some object property but depend on their context of "use".
Eg:
   *) a data changed in the GUI should be saved
   *) a data filled with a script shouldn't
etc...

To allow to control fine grain the state of a data is now set to non persitant
by default and should only be activated specifically (by the GUI, or the script)
if it is needed to save it.

This is a behavior change...but it is on a totally broken feature :)