[StentExp] CLEAN and UPDATE MultiBeamForceField code
Given the number of modifications, this is obviously not an atomic
commit. However, most of the modifications in MultiBeamForceField.inl
result from a reorganisation of the code, rather than from concrete
changes in the computation.
Such changes remain sparse in this commit, as the last major changes
in the algorithm correspond to revision 4d31c8f1 (March 2019).

Below is a list of the (minor) behaviour modifications introduced by the
current commit :
  - Most DEBUG traces and functions, which were commented, have been
deleted.
  - We created a new representation for tensor quantities, which we call
"vector notation". This notation is different from the Voigt notation in
that all the tensor's components are represented in a vector form
(leading to a 9-dimensional vector). Although this representation is
uselessly heavy for most tensor manipulations, it is naturally more
correct than the Voigt notation when dealing with inner products or
tensor products. The reason for this is that an inner product between
two tensors of order 2 cannot be accurately computed by the inner
product of two 6-dimensional vectors expressed in Voigt notation,
as half of the non-diagonal components is missing in the summation. The
most efficient way would be to store tensors in Voigt notation, and
define functions to accurately compute inner products between the
corresponding vectors, which would take into account all the tensor
components. This should be developed in future commits. In the meantime,
we mostly use the vector notation, which is more computationnaly costly,
but less likely to generate errors.
  - We suppressed the yield stress attribute in the BeamInfo structure,
as we consider that only one material is represented by one
MultiBeamForceField component. The yield stress of such material is
stored in the global Data _yieldStress.
  - As stated in introduction, the order of the methods in
MultiBeamForceField.inl is changed, to try to increase the readability
of the code.