[StentExp] CHANGE _virtualDisplacementMethod parameter
Resulting from an old version of the MultiBeamFEMForceField component,
the _virtualDisplacementMethod parameter is changed into
_usePrecomputedStiffness.
Basically, is does not alter much the parameter function in the
algorithm, but gives it a much more explicit name. Now, we use this
parameter to detrmine if the base stiffness matrix, used for elastic
deformation, should be initially computed using reduced integration, or
set up directly using a precomputed form (found in litterature).
Technically, this shouldn't change anything in the computation, as the
Gaussian reduced integration we perform is supposed to be exact in the
context of linear elasticity. However, small differences can occur due
to numerical computation of high values, therefore stressing the interest
of keeping both options.